feature/BWA-5-weather-app-api-service-refactor - Create Service provider and register services - Create services for each api and process - Refactor API Controller
