+++ src/app/Services/GeoapifyApi/Interfaces/GeoapifyApiServiceInterface.php
@@ -1,0 +1,10 @@
+<?php
+
+declare(strict_types=1);
+
+namespace App\Services\GeoapifyApi\Interfaces;
+
+interface GeoapifyApiServiceInterface
+{
+    public function getGeolocation(string $location): array;
+}
