+++ src/app/Services/OpenWeatherApi/Interfaces/OpenWeatherApiServiceInterface.php
@@ -1,0 +1,15 @@
+<?php
+
+declare(strict_types=1);
+
+namespace App\Services\OpenWeatherApi\Interfaces;
+
+use App\Enums\WeatherTypeEnum;
+use Illuminate\Http\Client\Response;
+
+interface OpenWeatherApiServiceInterface
+{
+    public function getWeatherForecast(float $lon, float $lat): Response;
+
+    public function getWeatherUpdate(float $lon, float $lat): Response;
+}
