+++ src/app/Services/UrlQueryStringBuilder/Interfaces/UrlQueryStringBuilderServiceInterface.php
@@ -1,0 +1,10 @@
+<?php
+
+declare(strict_types=1);
+
+namespace App\Services\UrlQueryStringBuilder\Interfaces;
+
+interface UrlQueryStringBuilderServiceInterface
+{
+    public function build(array $queryParam): string;
+}
